reorganized interface directory file structure into animal and habitat directories
