+++ interfaces/__init__.py
@@ -1,9 +1,9 @@
-from .terrestrial import ITerrestrial
+from .habitat.terrestrial import ITerrestrial
 from .animal.freshwater import IFreshwater
-from .aquatic import IAquatic
-from .flying import IFlying
-from .walking import IWalking
-from .swimming import ISwimming
+from .habitat.aquatic import IAquatic
+from interfaces.animal.flying import IFlying
+from interfaces.animal.walking import IWalking
+from interfaces.animal.swimming import ISwimming
 from .identifiable import Identifiable
 from .habitat.contains_animals import IContainsAnimals
 from .habitat.contains_plants import IContainsPlants
